backend/requirements.txt
Flask
Flask-RESTful
Flask-JWT-Extended
Flask-Limiter
Flask-mysqldb
Flask-Mail
pydantic
pydantic[email]
redis
python-dotenv